10373: create shared function for asyncHandleLockError;
